"changes 10-22 8:58"

+++ CS321_W5D2_BlogAPI.Core/Services/PostService.cs
@@ -19,11 +19,24 @@ namespace CS321_W5D2_BlogAPI.Core.Services
 
         public Post Add(Post newPost)
         {
-            // TODO: Prevent users from adding to a blog that isn't theirs
+            var currentUserId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(newPost.BlogId);
+            if(currentUserId==blog.UserId)
+            {
+                newPost.DatePublished = DateTime.Today;
+                return _postRepository.Add(newPost);
+            }
+            else
+            {
+                throw new ApplicationException("not your blog!");
+            }
+
+
+            // : Prevent users from adding to a blog that isn't theirs
             //     Use the _userService to get the current users id.
             //     You may have to retrieve the blog in order to check user id
-            // TODO: assign the current date to DatePublished
-            return _postRepository.Add(newPost);
+            // : assign the current date to DatePublished
+
         }
 
         public Post Get(int id)
@@ -44,14 +57,35 @@ namespace CS321_W5D2_BlogAPI.Core.Services
         public void Remove(int id)
         {
             var post = this.Get(id);
-            // TODO: prevent user from deleting from a blog that isn't theirs
-            _postRepository.Remove(id);
+            // prevent user from deleting from a blog that isn't theirs
+
+            var currentId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(post.BlogId);
+            if (currentId==blog.UserId)
+            {
+                _postRepository.Remove(id);
+            }
+            else
+            {
+                throw new Exception("you don't have permission to delete this post");
+            }
         }
 
         public Post Update(Post updatedPost)
         {
             // TODO: prevent user from updating a blog that isn't theirs
-            return _postRepository.Update(updatedPost);
+
+            var currentUserId = _userService.CurrentUserId;
+            var blog = _blogRepository.Get(updatedPost.BlogId);
+            if(currentUserId==blog.UserId)
+            {
+                return _postRepository.Update(updatedPost);
+            }
+            else
+            {
+                throw new Exception("this is not your blog!");
+            }
+            
         }
 
     }
